+++ src/Sources/Roosterkid.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace IlmLV\ProxyScraper\Sources;
+
+use IlmLV\ProxyScraper\Scrapers\TextListScraper;
+
+final class Roosterkid extends TextListScraper
+{
+    protected array $protocols = [
+        'https' => 'https://raw.githubusercontent.com/roosterkid/openproxylist/main/HTTPS_RAW.txt',
+        'socks4' => 'https://raw.githubusercontent.com/roosterkid/openproxylist/main/SOCKS4_RAW.txt',
+        'socks5' => 'https://raw.githubusercontent.com/roosterkid/openproxylist/main/SOCKS5_RAW.txt',
+    ];
+
+    public const SCHEDULE = '0 * * * *';
+}