+++ src/Validations/DomainsValidation.php
@@ -5,9 +5,10 @@ declare(strict_types=1);
 namespace IlmLV\ProxyScraper\Validations;
 
 use IlmLV\ProxyScraper\Validations\Domains\AbstractDomainValidation;
+use JsonSerializable;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
-class DomainsValidation implements \JsonSerializable, ValidationInterface
+class DomainsValidation implements JsonSerializable, ValidationInterface
 {
     use KeyedResultMap;
 
@@ -50,6 +51,8 @@ class DomainsValidation implements \JsonSerializable, ValidationInterface
 
     public function run(): self
     {
+        $this->validators = [];
+
         foreach ($this->validatorClasses as $validatorClass) {
             $validator = $validatorClass::make($this->client)->run();
             $this->validators[$validator::NAME] = $validator;